+++ packages/website/components/account/filesManager/filesManager.js
@@ -112,7 +112,7 @@ const FilesManager = ({ className, content, onFileUpload }) => {
   }, [fetchPinsDate, listPinned, isFetchingPinned, apiToken]);
   useEffect(() => {
     getTokens();
-  }, []);
+  }, [getTokens]);
 
   // Set displayed files based on tab selection: 'uploaded' or 'pinned'
   useEffect(() => {