+++ packages/website/modules/docs-theme/toc/toc.js
@@ -13,6 +13,7 @@ export default function Toc() {
   const [nestedHeadings, setNestedHeadings] = useState([]);
   const [updateKey, setUpdateKey] = useState(0);
   const activeHeadings = useRef({});
+  const lastActive = useRef('');
   let controller;
 
   const toggleClass = () => {
@@ -35,21 +36,31 @@ export default function Toc() {
     return str;
   };
 
-  const updateHeadings = (id) => {
+  const updateHeadings = id => {
+    let lastId = '';
     for (const element in activeHeadings.current) {
       if (activeHeadings.current[element]) {
         activeHeadings.current[element] = false;
+        lastId = element;
+        break;
       }
     }
-    activeHeadings.current[id] = true;
-    setUpdateKey(key => key + 1);
-  }
+    if (lastId !== id) {
+      activeHeadings.current[id] = true;
+      lastActive.current = lastId;
+      setUpdateKey(key => key + 1);
+    }
+  };
 
   const getNestedHeadings = headingElements => {
     const nestedHeadings = [];
     const headingIds = [];
+    const idCounts = {};
+
     headingElements.forEach((heading, index) => {
-      const id = string_to_slug(heading.innerText);
+      const innerTextSlug = string_to_slug(heading.innerText);
+      idCounts[innerTextSlug] = idCounts[innerTextSlug] ? idCounts[innerTextSlug] + 1 : 1;
+      const id = `${innerTextSlug}${idCounts[innerTextSlug] > 1 ? `-${idCounts[innerTextSlug]}` : ''}`;
       const title = heading.innerText;
       activeHeadings.current[id] = false;
       headingIds.push(id);
@@ -63,17 +74,21 @@ export default function Toc() {
       }
 
       // scroll effects / add active class
-      const exitScene = (e) => {
+      const exitScene = e => {
         if (e.scrollDirection === 'REVERSE' && index !== 0) {
           updateHeadings(headingIds[index - 1]);
         }
-      }
+      };
 
       const scene = new ScrollMagic.Scene({
         triggerElement: `#${id}`,
         triggerHook: 0.25,
-        duration: 200
-      }).on('enter', () => { updateHeadings(id) }).addTo(controller);
+        duration: 200,
+      })
+        .on('enter', () => {
+          updateHeadings(id);
+        })
+        .addTo(controller);
       scene.on('leave', exitScene).addTo(controller);
     });
     return nestedHeadings;
@@ -85,7 +100,11 @@ export default function Toc() {
         <li key={heading.id}>
           <a
             href={`#${heading.id}`}
-            className={clsx(activeHeadings.current[heading.id] ? 'active' : '')}>
+            className={clsx(
+              activeHeadings.current[heading.id] ? 'active' : '',
+              heading.id === lastActive.current ? 'heading-slide-out' : ''
+            )}
+          >
             {heading.title}
           </a>
           {heading.items.length > 0 && (
@@ -94,7 +113,11 @@ export default function Toc() {
                 <li key={child.id}>
                   <a
                     href={`#${child.id}`}
-                    className={clsx(activeHeadings.current[child.id] ? 'active' : '')}>
+                    className={clsx(
+                      activeHeadings.current[child.id] ? 'active' : '',
+                      child.id === lastActive.current ? 'heading-slide-out' : ''
+                    )}
+                  >
                     {child.title}
                   </a>
                 </li>
@@ -112,6 +135,16 @@ export default function Toc() {
     const newNestedHeadings = getNestedHeadings(headingElements);
     // @ts-ignore
     setNestedHeadings(newNestedHeadings);
+
+    // setTimeout needed to wait for the dom to be rendered
+    setTimeout(() => {
+      if (window.location.hash) {
+        const el = document.getElementById(window.location.hash.substring(1));
+        if (el) {
+          controller.scrollTo(window.location.hash);
+        }
+      }
+    });
   }, []);
 
   return (